requirements.txt
streamlit>=1.20
python-pptx>=0.6.21
Pillow>=9.0.0
opencv-python-headless>=4.5
numpy>=1.21
google-cloud-vision>=3.0.0
pytesseract>=0.3.10
